ast: rework ahb access logic to use common struct
Instead of shifting argv in hope the data is correct, use a common
arguments struct that can be easily implemented in a new command.

The logic of read or write must be done in the argument parsing of the
command. This reduces the excessive, and undocumented read/write call
logic to determine the correct argv values.

Signed-off-by: Tan Siewert <tan@siewert.io>